pyproject.toml
[tool.poetry]
name = "federatedopenwallet"
version = "0.1.0"
description = ""
authors = ["Jonas Legion <jturbeaux@pm.me>"]
license = "AGLv3"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.8"
django-solo = "2.1.0"
django-extensions = "3.2.3"
werkzeug = "^3.0.1"
django-browser-reload = "^1.11.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.4"
django-filter = "^23.5"
djangorestframework-api-key = "^3"
django-stdimage = "^6.0.1"
stripe = "^7.10.0"
coverage = "^7.4.0"
python-dotenv = "^1.0.0"
gunicorn = "^21.2.0"
cryptography = "^41.0.3"
channels = {extras = ["daphne"], version = "^4.0.0"}
django-htmx = "^1.16.0"
faker = "^22"
ipdb = "^0.13.13"
sentry-sdk = {extras = ["django"], version = "^1.39.1"}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
